+++ docs/changelog.rst
@@ -17,6 +17,9 @@ Modifications to existing flavors:
 
 - Fixed century bug with Kuwait Civil ID verification localflavor.kw.forms
   (`gh-195 <https://github.com/django/django-localflavor/pull/195>`_).
+-  Allow passing field name as first positional argument of IBANField.
+  (`gh-236 <https://github.com/django/django-localflavor/pull/236>`_).
+
 
 1.3   (2016-05-06)
 ------------------